# Fix #13654, improve performance of auto match for quotes