@suzaku 2022-07-16
BTW, I wonder why we use `_preceding_text_cache`, `re.compile` already has its own cache, doesn't it?