@Carreau 2022-08-30
Apologies for the delay, 

I think that conflicts with #13713 that I already merge, sorry about that. 
I'll see if I can port your improvement, unless you want to take a shot at it.